Refactor Remix to separate component
I'm realizing/discovering that the current implementation of this site is not meeting my needs very well anymore.  The changes in this commit represent a stop-gap measure to get the new project added in the meantime while I start to invest some time into investigating my options for setting up a headless CMS to handle the content side of the site.